# selector-max-specificity rule incorrectly reporting specificity