--- a/package.json
+++ b/package.json
@@ -54,7 +54,7 @@
     "postcss-selector-parser": "^2.0.0",
     "postcss-value-parser": "^3.1.1",
     "resolve-from": "^2.0.0",
-    "specificity": "^0.1.5",
+    "specificity": "^0.2.1",
     "string-width": "^1.0.1",
     "stylehacks": "^2.3.0",
     "sugarss": "^0.1.2",