The `selector-max-specificity` rule is incorrectly reporting that a selector with 11 elements has a higher specificity than a selector with a single classname. This is not an issue with the [specificity calculator](https://www.npmjs.com/package/specificity) module but in the way that this rule compares specificity as a base 10 number.

I have updated the unit tests and added a fix for this issue.

I recently updated the specificity calculator module to include a "compare" function, so I have used that function in this rule.

It doesn't appear to have had any negative performance impact.

Before:

```
MacBook:stylelint keegan$ npm run benchmark-rule -- selector-max-specificity "0,2,0"

> stylelint@6.6.0 benchmark-rule /Users/keegan/Projects/Tools/stylelint
> babel-node scripts/benchmark-rule.js "selector-max-specificity" "0,2,0"

Warnings: 900
Mean: 225.90502433333327 ms
Deviation: 67.87000886072683 ms
```

After:

```
MacBook:stylelint keegan$ npm run benchmark-rule -- selector-max-specificity "0,2,0"

> stylelint@6.6.0 benchmark-rule /Users/keegan/Projects/Tools/stylelint
> babel-node scripts/benchmark-rule.js "selector-max-specificity" "0,2,0"

Warnings: 900
Mean: 215.66833793103447 ms
Deviation: 84.2738416992734 ms
```