@jeddy3 2016-06-20
@keeganstreet Very nice! Looks great to me, but I'll wait for a 2nd pair of eyes before merging.

In the meantime, could you please add a "Fixed: `selector-max-specificity` no longer reports that a selector with 11 elements has a higher specificity than a selector with a single classname" item to the changelog please.

Thank you for creating the `specificity` module btw. It's pretty awesome :)