@evilebottnawi 2016-06-20
@jeddy3 LGTM, never would have thought that 101 type selectors have lower specificity than a one class selector.