@jeddy3 2016-06-20
I added the changelog item for you in https://github.com/stylelint/stylelint/commit/42fbc13ae913352cbd2924dfe67defb76598bd36

So we can get this into the upcoming release.