@keeganstreet 2016-06-20
Thanks @jeddy, @evilebottnawi & @davidtheclark! Great work on stylelint. Its a pleasure submitting a PR to a well organised project like this, with the docs, unit tests, performance tests, etc.

Thanks for updating the changelog @jeddy3. I was asleep while this discussion was going on - different time zones.

@evilebottnawi yeah its interesting that a single classname has a higher specificity than an infinite number of element names - not that you would probably ever want to have selectors with that many parts, but hey, that's what the `selector-max-compound-selectors` rule is for :-) There actually used to be bugs in most major browsers where 256 elements would have a higher specificity than a classname. Its been fixed in Chrome and Safari but still exists in Firefox. https://www.thecssninja.com/css/extreme-specificity