@davidtheclark 2016-06-21
> Great work on stylelint. Its a pleasure submitting a PR to a well organised project like this, with the docs, unit tests, performance tests, etc.

That's really good to hear, thanks!