@jeddy3 2016-06-21
>  I was asleep while this discussion was going on - different time zones.

That happens all the time to us :) The core contributors are spread across the USA, UK, Russia and Australia.

> That's really good to hear, thanks!

Agreed! Thanks :)