@evilebottnawi 2016-06-21
@keeganstreet hm, very intresting, the more I work on stylelint, the more I learn about CSS :smile: