@ntwb 2016-06-22
>  hm, very intresting, the more I work on stylelint, the more I learn about CSS

Me too 👍